+++ CHANGELOG.md
@@ -31,6 +31,13 @@
   existence of `syncBookmarks`.
   ([#850](https://github.com/mozilla/application-services/issues/850))
 
+## FxA
+
+### What's New
+
+- New methods `getManageAccountURL` and `getManageDevicesURL` have been added,
+  which the application can use to direct the user to manage their account on the web.
+
 # v0.25.2 (_2018-04-11_)
 
 [Full Changelog](https://github.com/mozilla/application-services/compare/v0.24.0...v0.25.2)